# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GIC Take-Home Backend",
    "version": "0.1.0"
  },
  "paths": {
    "/cafes": {
      "get": {
        "tags": [
          "cafes"
        ],
        "summary": "List Cafes Endpoint",
        "description": "Return cafes with optional normalized location filtering.",
        "operationId": "list_cafes_endpoint_cafes_get",
        "parameters": [
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CafeListItem"
                  },
                  "title": "Response List Cafes Endpoint Cafes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "cafes"
        ],
        "summary": "Create Cafe Endpoint",
        "description": "Create one cafe.",
        "operationId": "create_cafe_endpoint_cafes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CafeWriteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CafeWriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cafes/{cafe_id}": {
      "get": {
        "tags": [
          "cafes"
        ],
        "summary": "Get Cafe Detail Endpoint",
        "description": "Return editable detail fields for one cafe.",
        "operationId": "get_cafe_detail_endpoint_cafes__cafe_id__get",
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cafe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CafeDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "cafes"
        ],
        "summary": "Update Cafe Endpoint",
        "description": "Replace the editable fields for one cafe.",
        "operationId": "update_cafe_endpoint_cafes__cafe_id__put",
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cafe Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CafeWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CafeWriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cafes"
        ],
        "summary": "Delete Cafe Endpoint",
        "description": "Delete one cafe and its currently assigned employees.",
        "operationId": "delete_cafe_endpoint_cafes__cafe_id__delete",
        "parameters": [
          {
            "name": "cafe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cafe Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/employees": {
      "get": {
        "tags": [
          "employees"
        ],
        "summary": "List Employees Endpoint",
        "description": "Return employees with an optional filter for their current cafe.",
        "operationId": "list_employees_endpoint_employees_get",
        "parameters": [
          {
            "name": "cafe",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cafe"
            }
          },
          {
            "name": "cafe_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cafe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EmployeeListItem"
                  },
                  "title": "Response List Employees Endpoint Employees Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "employees"
        ],
        "summary": "Create Employee Endpoint",
        "description": "Create one employee with an initial cafe assignment.",
        "operationId": "create_employee_endpoint_employees_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeWriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/employees/{employee_id}": {
      "get": {
        "tags": [
          "employees"
        ],
        "summary": "Get Employee Detail Endpoint",
        "description": "Return editable detail fields for one employee.",
        "operationId": "get_employee_detail_endpoint_employees__employee_id__get",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Employee Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "employees"
        ],
        "summary": "Update Employee Endpoint",
        "description": "Replace the editable fields for one employee and apply assignment changes.",
        "operationId": "update_employee_endpoint_employees__employee_id__put",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Employee Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmployeeWriteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmployeeWriteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "employees"
        ],
        "summary": "Delete Employee Endpoint",
        "description": "Delete one employee and all assignment rows tied to them.",
        "operationId": "delete_employee_endpoint_employees__employee_id__delete",
        "parameters": [
          {
            "name": "employee_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Employee Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Healthcheck",
        "operationId": "healthcheck_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Healthcheck Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Readiness Check",
        "operationId": "readiness_check_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CafeDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "location": {
            "type": "string",
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "logo_url",
          "location"
        ],
        "title": "CafeDetail",
        "description": "Represents the editable cafe fields used by the detail endpoint."
      },
      "CafeListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "location": {
            "type": "string",
            "title": "Location"
          },
          "employees": {
            "type": "integer",
            "title": "Employees"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "logo_url",
          "location",
          "employees"
        ],
        "title": "CafeListItem",
        "description": "Represents one cafe row returned by the list endpoint."
      },
      "CafeWriteRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "location": {
            "type": "string",
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "location"
        ],
        "title": "CafeWriteRequest",
        "description": "Represents the request payload shared by cafe create and update operations."
      },
      "CafeWriteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "logo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Url"
          },
          "location": {
            "type": "string",
            "title": "Location"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "logo_url",
          "location"
        ],
        "title": "CafeWriteResponse",
        "description": "Represents the persisted cafe payload returned after writes."
      },
      "EmployeeCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email_address": {
            "type": "string",
            "title": "Email Address"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "cafe_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cafe Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email_address",
          "phone_number",
          "gender"
        ],
        "title": "EmployeeCreateRequest",
        "description": "Represents the create payload for employee writes."
      },
      "EmployeeDetail": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email_address": {
            "type": "string",
            "title": "Email Address"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "cafe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cafe"
          },
          "cafe_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cafe Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email_address",
          "phone_number",
          "gender",
          "cafe",
          "cafe_id"
        ],
        "title": "EmployeeDetail",
        "description": "Represents the editable employee fields returned by the detail endpoint."
      },
      "EmployeeListItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email_address": {
            "type": "string",
            "title": "Email Address"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "days_worked": {
            "type": "integer",
            "title": "Days Worked"
          },
          "cafe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cafe"
          },
          "cafe_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cafe Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email_address",
          "phone_number",
          "gender",
          "days_worked",
          "cafe",
          "cafe_id"
        ],
        "title": "EmployeeListItem",
        "description": "Represents one employee row returned by the list endpoint."
      },
      "EmployeeWriteRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email_address": {
            "type": "string",
            "title": "Email Address"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "cafe_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cafe Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email_address",
          "phone_number",
          "gender"
        ],
        "title": "EmployeeWriteRequest",
        "description": "Represents the request payload shared by employee create and update operations."
      },
      "EmployeeWriteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email_address": {
            "type": "string",
            "title": "Email Address"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "days_worked": {
            "type": "integer",
            "title": "Days Worked"
          },
          "cafe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cafe"
          },
          "cafe_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cafe Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email_address",
          "phone_number",
          "gender",
          "days_worked",
          "cafe",
          "cafe_id"
        ],
        "title": "EmployeeWriteResponse",
        "description": "Represents the saved employee payload returned after writes."
      },
      "Gender": {
        "type": "string",
        "enum": [
          "Male",
          "Female"
        ],
        "title": "Gender",
        "description": "Canonical gender values defined by the assignment contract."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}